Add loading files
