+++ src/app/[locale]/(app)/explore/_components/list-places-of-category.tsx
@@ -1,6 +1,7 @@
 'use client'
 
 import { Card, CardBody } from '@nextui-org/card'
+import { Skeleton } from '@nextui-org/react'
 import { IconChevronRight, IconPlus } from '@tabler/icons-react'
 import { useTranslations } from 'next-intl'
 import { FC } from 'react'
@@ -92,6 +93,29 @@ export const ListPlacesOfCategory: FC<{
           </Link>
         </li>
       </ul>
+    </div>
+  )
+}
+
+export const ListPlacesOfCategorySkeleton: FC = () => {
+  return (
+    <div className="space-y-2 py-4">
+      <div className="flex items-center gap-4 px-4">
+        <Skeleton className="h-10 w-10 rounded-full" />
+        <Skeleton className="h-6 w-full max-w-32 rounded-full" />
+        <div className="grow" />
+        <Skeleton className="h-3 w-full max-w-16 rounded-full" />
+        <IconChevronRight stroke={1.5} className="-ml-4 text-stone-300" />
+      </div>
+
+      <div className="space-x-4 overflow-x-hidden whitespace-nowrap pl-4 scrollbar-hide">
+        {[...Array(3)].map((_, i) => (
+          <div className="inline-block w-[calc(50vw-3rem)] min-w-32 max-w-72 space-y-2">
+            <Skeleton key={i} className="aspect-[4/3] w-full rounded-xl" />
+            <Skeleton className="mx-auto h-4 w-full max-w-28 rounded-full" />
+          </div>
+        ))}
+      </div>
     </div>
   )
 }
