+++ src/app/[locale]/(app)/explore/places/loading.tsx
@@ -1,0 +1,17 @@
+import { FC } from 'react'
+import { CategoriesGridSkeleton } from '../_components/categories-grid'
+import { ListPlacesOfCategorySkeleton } from '../_components/list-places-of-category'
+
+const Loading: FC = () => {
+  return (
+    <>
+      <CategoriesGridSkeleton />
+
+      {[...Array(5)].map((_, i) => (
+        <ListPlacesOfCategorySkeleton key={i} />
+      ))}
+    </>
+  )
+}
+
+export default Loading
